feat: 初步完成文章前端控制器

+++ personal-blog-admin/src/main/java/com/gewuyou/blog/admin/service/impl/UserAuthServiceImpl.java
@@ -295,7 +295,7 @@ public class UserAuthServiceImpl extends ServiceImpl<UserAuthMapper, UserAuth> i
         }
         Page<UserAdminDTO> page = new Page<>(PageUtil.getCurrent(), PageUtil.getSize());
         List<UserAdminDTO> userAdminDTOS = baseMapper.listUsers(page, conditionVO);
-        return new PageResultDTO<>(userAdminDTOS, count);
+        return new PageResultDTO<>(userAdminDTOS, count.longValue());
     }
 
     /**
